🏗️ Architect: Modularized plotting dispatcher into wrappers
Refactored the 434-line `apts/plot.py` 'God Object' dispatcher into a modular package structure within `apts/plotting/`.

Key changes:
- Created `apts/plotting/dispatcher.py` to house the `Plotter` and `NullPlotter` classes (reduced to 149 lines).
- Extracted plotting wrapper functions into specialized modules: `apts/plotting/wrappers/objects.py` and `apts/plotting/wrappers/weather.py`.
- Converted `apts/plot.py` into a thin, backward-compatible shim.
- Centralized the plotting API in `apts/plotting/__init__.py`.
- Updated internal references in `apts/observations/` to use `apts.plotting`.
- Implemented robust centroid column detection in `apts/utils/fits_analyzer.py` to support both `photutils < 3.0.0` (`xcentroid`) and `>= 3.0.0` (`x_centroid`) naming conventions.
- Updated unit test mocks to align with the new package structure.

Verified all 719 unit tests pass.
